--- header-range/Cargo.toml
@@ -1,18 +1,0 @@
-[workspace]
-[package]
-version = "0.1.0"
-name = "sp1-vectorx-header-range-program"
-edition = "2021"
-
-[dependencies]
-sp1-vectorx-primitives = { path = "../primitives" }
-sp1-zkvm = { git = "https://github.com/succinctlabs/sp1.git", branch = "dev" }
-ed25519-consensus = { version = "2", default-features = false }
-blake2 = "0.10.6"
-sha2 = "0.10.8"
-alloy-sol-types = "0.7.4"
-alloy-primitives = "0.7.4"
-
-[patch.crates-io]
-ed25519-consensus = { git = "https://github.com/sp1-patches/ed25519-consensus", branch = "patch-v2.1.0" }
-sha2-v0-10-8 = { git = "https://github.com/sp1-patches/RustCrypto-hashes", package = "sha2", branch = "patch-v0.10.8" }
